% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeQuantileSupportFunctions.R
\name{index_collapser}
\alias{index_collapser}
\title{Function to turn indeces from p-dimensions into one set of unique indeces}
\usage{
index_collapser(nbins, indeces)
}
\arguments{
\item{nbins}{number of bins in each dimension}

\item{indeces}{p-dimensional indeces for bin}
}
\value{
integer index unique to p-dim bin
}
\description{
take p-dim indeces and convert fast/slow running indeces
}
\examples{
index_collapser(nbins=c(2,2,2),indeces=c(1,1,1))
index_collapser(nbins=c(2,2,2),indeces=c(1,1,2))
index_collapser(nbins=c(2,2,2),indeces=c(1,2,1))
index_collapser(nbins=c(2,2,2),indeces=c(1,2,2))
index_collapser(nbins=c(2,2,2),indeces=c(2,1,1))
index_collapser(nbins=c(2,2,2),indeces=c(2,1,2))
index_collapser(nbins=c(2,2,2),indeces=c(2,2,1))
index_collapser(nbins=c(2,2,2),indeces=c(2,2,2))
}
