% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeQuantileBinning.R
\name{iqbin}
\alias{iqbin}
\title{Iterative Quantile Binning}
\usage{
iqbin(data, bin_cols, nbins, jit = rep(0, length(bin_cols)),
  output = "data")
}
\arguments{
\item{data}{data frame to be binned (will coerce matrix or tibble to simple data frame)}

\item{bin_cols}{vector of column names of variables to iteratively bin, ordered first to last}

\item{nbins}{vector of number of bins per step of iterative binning, ordered first to last}

\item{jit}{vector of margins for uniform jitter to each dimension to create seperability of tied obs due to finite precision}

\item{output}{Output Structure: "data" for just the binned data,"definition" for a list of bin centers and boundaries, or "both" for list containing both data and definition}
}
\value{
output list containing values specified by user
}
\description{
used for binning the numeric values by empirical quantile
}
\examples{
iqbin(data=iris, bin_cols=c("Sepal.Length","Sepal.Width","Petal.Width"),
                    nbins=c(3,5,2), output="both",jit=rep(0.001,3))
iqbin(data=iris, bin_cols=c("Sepal.Length","Sepal.Width","Petal.Width"),
                    nbins=c(3,5,2), output="data")
}
\seealso{
Other iterative quantile binning functions: \code{\link{iqbin_assign}},
  \code{\link{iqbin_stretch}}
}
