% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeQuantilePlotting.R
\name{iqbin_plot_2d}
\alias{iqbin_plot_2d}
\title{Plotting 2-dimensional iterative quantile bins}
\usage{
iqbin_plot_2d(iq_obj)
}
\arguments{
\item{iq_obj}{iterative quantile object: either from iqbin or iqnn function with data and definition}
}
\value{
create ggplot2 object with iq bins displayed
}
\description{
used for binning the numeric values by empirical quantile
}
\examples{
iq_obj <- iqbin(data=iris, bin_cols=c("Sepal.Length","Sepal.Width"), nbins=c(5,3), output="both",jit=rep(0.1,2))
iqbin_plot_2d(iq_obj)

iq_obj <- iqbin(data=iris, bin_cols=c("Sepal.Length","Sepal.Width"), nbins=c(5,3), output="both")
iqbin_plot_2d(iq_obj)

iq_obj <- iqnn(data=iris, y="Species", mod_type="class", bin_cols=c("Sepal.Length","Sepal.Width"), nbins=c(5,3), jit=rep(0.001,2))
iqbin_plot_2d(iq_obj)

iq_obj <- iqnn(data=iris, y="Petal.Length", mod_type="reg", bin_cols=c("Sepal.Length","Sepal.Width"), nbins=c(5,3), jit=rep(0.001,2))
iqbin_plot_2d(iq_obj)
}
