% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeQuantilePlotting.R
\name{iqbin_plot_3d}
\alias{iqbin_plot_3d}
\title{Plotting 3-dimensional iterative quantile bins}
\usage{
iqbin_plot_3d(iq_obj, round_digits = 2)
}
\arguments{
\item{iq_obj}{iterative quantile object: either from iqbin or iqnn function with data and definition}
}
\value{
create ggplot2 object with iq bins displayed
}
\description{
Plotting 3-dimensional iterative quantile bins using faceting
}
\examples{
iq_obj <- iqbin(data=iris, bin_cols=c("Sepal.Length","Sepal.Width","Petal.Width"),
                nbins=c(2,5,3), output="both",jit=rep(0.1,3))
iqbin_plot_3d(iq_obj)

iq_obj <- iqnn(data=iris, y="Species", mod_type="class", bin_cols=c("Sepal.Length","Sepal.Width","Petal.Width"),
               nbins=c(2,5,3), jit=rep(0.001,3))
iqbin_plot_3d(iq_obj)

iq_obj <- iqnn(data=iris, y="Petal.Length", mod_type="reg", bin_cols=c("Sepal.Length","Sepal.Width","Petal.Width"),
                nbins=c(2,5,3), jit=rep(0.001,3))
iqbin_plot_3d(iq_obj)
}
