% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeQuantileBinning.R
\name{iqbin_stretch}
\alias{iqbin_stretch}
\title{Adding Tolerance Buffer to outermost bins from Iterative Quantile Binning}
\usage{
iqbin_stretch(iq_def, tol)
}
\arguments{
\item{iq_def}{iterative quantile binning definition list}

\item{tol}{vector of tolerance values to stretch each dimension for future binning}
}
\value{
updated binning definition with bins extended by tolerance values
}
\description{
New observations selected from the same population as the data used to build bin definitions may fall just outside the bins. If we wish to include nearby values we can either allow outer bins to be extended (this function) or to leave the outer bins unbounded.
}
