% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeQuantileNearestNeighbors.R
\name{iqnn}
\alias{iqnn}
\title{Iterative Quantile Binned Nearest Neighbors}
\usage{
iqnn(data, y, mod_type = "reg", bin_cols, nbins, jit = rep(0,
  length(bin_cols)), stretch = FALSE, tol = rep(0, length(bin_cols)))
}
\arguments{
\item{data}{Data frame containing the response variable and numeric input variables from the training data}

\item{y}{Name of response variable column}

\item{mod_type}{Depends on response variables type: "reg" creates iqnn-regression for predicting numeric values, "class" creates iqnn-classifier for predicting categorical values}

\item{bin_cols}{vector of column names of variables to iteratively bin, ordered first to last}

\item{nbins}{vector of number of bins per step of iterative binning, ordered first to last}

\item{jit}{vector of margins for uniform jitter to each dimension to create seperability of tied obs due to finite precision}

\item{stretch}{TRUE/FALSE if will bins be given tolerance buffer}

\item{tol}{vector of tolerance values to stretch each dimension for future binning}
}
\value{
list containing binned training data, binning definition, and bin statistics
}
\description{
Function for creating iterative quantile nearest neighbors model. Bin the training data, then store the binning definitions and bin statistics to be used to estimate for future testing data.
}
\examples{
iqnn_mod <- iqnn(data=iris, y="Species", mod_type="class", bin_cols=c("Sepal.Length","Sepal.Width","Petal.Width"),
                 nbins=c(3,5,2), jit=rep(0.001,3), tol = rep(0.001,3))
str(iqnn_mod)                
}
