% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeQuantileNearestNeighbors.R
\name{iqnn_cv_predict}
\alias{iqnn_cv_predict}
\title{Cross Validated predictions for iqnn models}
\usage{
iqnn_cv_predict(data, y, mod_type = "reg", bin_cols, nbins, jit = rep(0,
  length(bin_cols)), stretch = FALSE, tol = rep(0, length(bin_cols)),
  strict = FALSE, cv_k = 10)
}
\arguments{
\item{data}{Data frame containing the response variable and numeric input variables from the training data}

\item{y}{Name of response variable column}

\item{mod_type}{Depends on response variables type: "reg" creates iqnn-regression for predicting numeric values, "class" creates iqnn-classifier for predicting categorical values}

\item{bin_cols}{vector of column names of variables to iteratively bin, ordered first to last}

\item{nbins}{vector of number of bins per step of iterative binning, ordered first to last}

\item{jit}{vector of margins for uniform jitter to each dimension to create seperability of tied obs due to finite precision}

\item{stretch}{TRUE/FALSE if will bins be given tolerance buffer}

\item{tol}{vector of tolerance values to stretch each dimension for future binning}

\item{strict}{TRUE/FALSE: If TRUE Observations must fall within existing bins to be assigned; if FALSE the outer bins in each dimension are unbounded to allow outlying values to be assigned.}

\item{cv_k}{integer specifying number of folds}
}
\value{
cross validated predicted responses for all observations in data
}
\description{
Cross-validate an iqnn specification using k-fold scheme on given data
}
