% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeQuantileSupportFunctions.R
\name{majority_vote}
\alias{majority_vote}
\title{Simple Majority Vote Counter}
\usage{
majority_vote(votes)
}
\arguments{
\item{votes}{character or factor vector}
}
\description{
Identify the maximum vote earners, then randomly pick winner if there is a tie to break
}
