% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeQuantileSupportFunctions.R
\name{make_cv_cohorts}
\alias{make_cv_cohorts}
\title{CV cohort additions}
\usage{
make_cv_cohorts(dat, cv_K)
}
\arguments{
\item{dat}{data frame for cross validation}

\item{cv_K}{number of folds needed in indexing}
}
\value{
Vector of fold indeces
}
\description{
Create set of indeces for tracking observations to cv folds. Creates equally balanced folds
}
\examples{
make_cv_cohorts(iris,cv_K=10)
}
