% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeQuantileSupportFunctions.R
\name{make_nbins_list}
\alias{make_nbins_list}
\title{Function to create list of nbins vectors to put into tuning iqnn}
\usage{
make_nbins_list(nbin_range, p)
}
\arguments{
\item{nbin_range}{positive integer vector containing lower and upper bounds on number of bins in each dimension}

\item{p}{number of binning dimensions}
}
\value{
list of nbins vectors
}
\description{
create a list of nbins vectors, use progression that increases number of bins in each dimension while always staying balanced between dimensions
}
\examples{
make_nbins_list(c(2,3),3)
}
