% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeQuantileSupportFunctions.R
\name{quant_bin_1d}
\alias{quant_bin_1d}
\title{One-Dimensional Empirical Quantile Binning}
\usage{
quant_bin_1d(xs, nbin, output = "data", jit = 0)
}
\arguments{
\item{xs}{Numeric vector of values to be binned}

\item{nbin}{An integer defining the number of bins to partion the xs into}

\item{output}{Output Structure: "data" for just the binned data,"definition" for a list of bin centers and boundaries, or "both" for list containing both data and definition}

\item{jit}{non-negative value to specify a random uniform jitter to the observed values prior to partitioning by quantiles.}
}
\value{
output as specified
}
\description{
used for binning the numeric values by empirical quantile
}
