% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeQuantileSupportFunctions.R
\name{round_df}
\alias{round_df}
\title{Create data frame with rounded number of trailing digits}
\usage{
round_df(x, digits = 2)
}
\arguments{
\item{x}{data frame}

\item{digits}{number of digits to round}
}
\value{
data frame with rounded numeric variables
}
\description{
Based on https://gist.github.com/avsmith/e6f4f654451da139230b to round all numeric variables
}
