% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeQuantilePlotting.R
\name{iqbin_plot_2d}
\alias{iqbin_plot_2d}
\title{Plotting 2-dimensional iterative quantile bins}
\usage{
iqbin_plot_2d(iq_obj)
}
\arguments{
\item{iq_obj}{iterative quantile object: either from iqbin or iqnn function with data and definition}
}
\value{
create ggplot2 object with iq bins displayed
}
\description{
used for binning the numeric values by empirical quantile
}
\examples{
2+2
}
