% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeQuantileBinning.R
\name{iqbin_stretch}
\alias{iqbin_stretch}
\title{Stretching to Add Tolerance Buffer to outermost Iterative Quantile Bins}
\usage{
iqbin_stretch(iq_def, tol)
}
\arguments{
\item{iq_def}{iterative quantile binning definition list from \code{\link{iqbin}} function}

\item{tol}{vector of tolerance values to stretch each dimension for future binning}
}
\value{
updated binning definition list with outermost bin boundaries extended by tolerance values
}
\description{
We may wish to include values just outside the constructed bins in future applications. \code{iqbin_stretch} redefines the outermost bin in each dimension from a definition created by \code{\link{iqbin}}
}
\examples{
iq_def <- iqbin(data=iris, bin_cols=c("Sepal.Length","Sepal.Width","Petal.Width"),
                             nbins=c(3,2,2), output="both")
stretch_iq_def <- iqbin_stretch(iq_def, tol = c(1,1,1))
iq_def$bin_def$bin_bounds
stretch_iq_def$bin_def$bin_bounds
}
\seealso{
Other iterative quantile binning functions: \code{\link{iqbin_assign}},
  \code{\link{iqbin}}
}
