% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeQuantileSupportFunctions.R
\name{make_bin_list}
\alias{make_bin_list}
\title{Convert bin bounds data frame into R-tree structure using nested lists}
\usage{
make_bin_list(bin_bounds, nbins)
}
\arguments{
\item{bin_bounds}{Data frame of bin bounds; one row per bin, columns for lower and upper bound on each dimension}

\item{nbins}{number of bins in each dimension}
}
\value{
R-tree nested list of bin boundaries
}
\description{
Convert bin bounds data frame into R-tree structure using nested lists
}
\examples{
nbins=c(5,4,3,2)
iq_def <- iqbin(data=iris, bin_cols=c("Sepal.Length","Sepal.Width","Petal.Length","Petal.Width"),
                    nbins=nbins, output="definition",jit=rep(0.001,4))
iq_def$bin_bounds
make_bin_list(bin_bounds=iq_def$bin_bounds,nbins=nbins)
}
